Change product ID type from long to int

+++ HibernateAssesment/src/main/java/com/cap/hibernate/HibernateAssesment/DeleteProduct.java
@@ -5,10 +5,11 @@ import org.hibernate.Session;
 import Entity.Product;
 
 public class DeleteProduct {
-	public void deleteProduct(Session session, long id) {
+	public void deleteProduct(Session session, int id) {
 		RetriveFromDatabase retrive = new RetriveFromDatabase();
 		Product product = retrive.retriveProduct(session, id);
 		session.delete(product);
 		System.out.println("Product deleted successfully");
 	}
 }
+
